requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
praw==7.1.0
prawcore==1.5.0
python-decouple==3.4
requests==2.25.1
six==1.15.0
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
